tests/e2e: move tcp test to own file and test permissive mode
Moves the TCP traffic test to its own file and adds a scenario
to test traffic in permissive mode.

Part of #1521
